plan,exec: carry the combine cxl body program on PlanNode::Combine (#621)
The combine `cxl:` body program now travels on the node like
`PlanTransformPayload.typed` / `Aggregation.compiled`, instead of being
fetched from the bare-name-keyed `CompileArtifacts.typed` side-table.

- `PlanNode::Combine` gains a `typed: Option<Arc<TypedProgram>>` field,
  populated at lowering from `artifacts.typed.get(name)` (mirroring the
  Transform arm).
- N-ary decomposition reads the body program off the original node and
  stamps it onto the final chain step; non-final steps are body-less and
  carry `None`. The flat-map remove/re-insert of the body is dropped.
- The executor's four combine_dispatch body-program reads now read the
  program off the destructured node rather than re-fetching by name.

Downstream lineage walks `dag().graph` and each `body_of().graph` and
reads the program off the node uniformly for top-level and body combines,
so a Combine is no longer a lineage dead end. This removes the interim
BoundBody accessor approach, which solved the same goal off-node.

The bare-name lookup at lowering remains collision-exposed exactly like
Transform's; scope-correct keying is tracked as a follow-up.
