+++ crates/clinker-plan/src/config/pipeline.rs
@@ -3588,6 +3588,14 @@ pub(crate) fn lower_node_to_plan_node(
                 .get(name)
                 .cloned()
                 .unwrap_or_else(|| Arc::new(std::collections::HashMap::new()));
+            // Carry the typed `cxl:` body program on the node, mirroring the
+            // Transform arm's `artifacts.typed.get(name)` read into
+            // `PlanTransformPayload.typed`. Body and top-level combines both
+            // flow through this arm; the bare-name lookup is collision-exposed
+            // exactly like Transform's (scope-correct keying is a later phase).
+            // `None` is legitimate here — a body-less combine (e.g.
+            // `match: collect`) carries no program.
+            let typed = artifacts.typed.get(name).cloned();
             Some(crate::plan::execution::PlanNode::Combine {
                 name: name.to_string(),
                 span,
@@ -3602,6 +3610,7 @@ pub(crate) fn lower_node_to_plan_node(
                 decomposed_from: None,
                 output_schema: schema_from_bound(name),
                 resolved_column_map,
+                typed,
             })
         }
         // Reshape lowers to PlanNode::Reshape carrying the parsed config
